requirements.txt
Babel==2.7.0
Django==2.2.2
django-environ==0.4.5
django-phonenumber-field==3.0.1
djangorestframework==3.9.4
gcloud==0.18.3
googleapis-common-protos==1.6.0
httplib2==0.13.0
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauth2client==4.1.3
phonenumbers==8.10.13
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
